+++ src/constants/constants.js
@@ -210,17 +210,76 @@ export const ACTIONS = {
 export const OPTIONS = {
   TEMPORARY_LOCATION: 'TEMPORARY_LOCATION',
   PERMANENT_LOCATION: 'PERMANENT_LOCATION',
+  STATUS: 'STATUS',
 };
 
 export const ITEMS_OPTIONS = [
+  {
+    value: 'TEMPORARY_LOCATION',
+    label: 'ui-bulk-edit.layer.options.temporary',
+    disabled: false,
+  },
   {
     value: 'PERMANENT_LOCATION',
     label: 'ui-bulk-edit.layer.options.permanent',
     disabled: false,
   },
   {
-    value: 'TEMPORARY_LOCATION',
-    label: 'ui-bulk-edit.layer.options.temporary',
+    value: 'STATUS',
+    label: 'ui-bulk-edit.layer.options.statusLabel',
+    disabled: false,
+  },
+];
+
+export const ITEM_STATUS_OPTIONS = [
+  {
+    value: '',
+    label: 'ui-bulk-edit.layer.options.placeholder',
+    disabled: true,
+  },
+  {
+    value: 'Available',
+    label: 'ui-bulk-edit.layer.options.available',
+    disabled: false,
+  },
+  {
+    value: 'Withdrawn',
+    label: 'ui-bulk-edit.layer.options.withdrawn',
+    disabled: false,
+  },
+  {
+    value: 'Missing',
+    label: 'ui-bulk-edit.layer.options.missing',
+    disabled: false,
+  },
+  {
+    value: 'In process (non-requestable)',
+    label: 'ui-bulk-edit.layer.options.non-requestable',
+    disabled: false,
+  },
+  {
+    value: 'Intellectual item"',
+    label: 'ui-bulk-edit.layer.options.intellectual',
+    disabled: false,
+  },
+  {
+    value: 'Long missing',
+    label: 'ui-bulk-edit.layer.options.longMissing',
+    disabled: false,
+  },
+  {
+    value: 'Restricted',
+    label: 'ui-bulk-edit.layer.options.restricted',
+    disabled: false,
+  },
+  {
+    value: 'Unavailable',
+    label: 'ui-bulk-edit.layer.options.unavailable',
+    disabled: false,
+  },
+  {
+    value: 'Unknown',
+    label: 'ui-bulk-edit.layer.options.unknown',
     disabled: false,
   },
 ];